Add definitions for SOCKET_INTERCEPT plugins
A socket intercept plugin can intercept and alter the processing of
all communication on the clear/Internet/underlying network for an
OpenVPN process which would otherwise be in UDP mode, by replacing
the native socket used for that communication with a virtual socket
mediated by the plugin. A new plugin type is added for this, and a
new function to be exported by the plugin to allow retrieving the
vtable for the virtual socket implementation.