Add skeletal and munging code for obfs-test plugin
This plugin will provide a basic example for use of the socket
intercept interface to transform data on the wire to work around
unwanted network filtering. The munging functions provided do a
simple transformation on the first bytes of data packets, as well
as changing the port number.

The networking functions themselves are not in this commit; they
will be in platform-specific commits. Thus, this plugin does not
actually work in this commit.