Add Windows virtual socket code for obfs-test plugin
This implements the virtual socket functions in terms of Winsock,
with the munging of data and addresses included. The io_slot
implementation that translates between overlapped I/O and the
readiness-polling interface used by the main event loop is similar
to the mechanism used in OpenVPN core. This plugin should now have
support for Windows platforms.