transport: introduce tranport API plugin codebase
This API will allow plugins to provide virtual socket implementations
with their own table of functions for operations on these sockets,
with limited but functional integration with the main OpenVPN event
loop.

This new mechanism can be used to implement alternative transport
protocols via external plugins, without interfering with the OpenVPN
core codebase.

A sample user of this API might be an obsufcation schema which
manipulates the traffic before sending/receiving it.

Signed-off-by: Robin Tarsiger <rtt@dasyatidae.com>
[antonio@openvpn.net: refactored commits, restyled code]
[gianmarco@mandelbit.com: rebased, adapted to multisocket]