Lay the foundation to support both old and new versions of DCO
Added a new abstraction layer based on the `struct dco_ops`,
a structure of function pointers that ensures the appropriate
APIs at runtime are used for DCO communication between kernel
and user space, selection is based on the available module's
family name.

The old version, referred to as "ovpn_dco_v2", and its related
commands and functions, will be distinguished with the "v2" suffix.

Signed-off-by: Gianmarco De Gregori <gianmarco@mandelbit.com>