route: extended logic to omit gateway when unnecessary
Extracted and extended the logic behind 'gateway_needed' both in
add_route() and add_route_ipv6(). Other than checking the dev-type,
special routes and if the gateway is on-link,
- set gateway_needed to true if the vpn instance is a multipoint server
  and DCO is enabled.
- set gateway_needed to false if the gateway is in the vpn subnet.
Additionally, extended support for these checks and conditions to DARWIN
and BSD-based operating systems.

These changes ensure that the gateway is only included when necessary,
optimizing route configuration and potentially reducing redundant route entries.

Change-Id: I3dbd61fef20d683734b55b73e40afe56c3b733e1
Signed-off-by: Marco Baffo <marco@mandelbit.com>