Add support for HAProxy's PROXY protocol
The PROXY protocol is a simple protocol that can be used to
communicate connection information such as a client's address across
multiple layers of NAT or TCP proxies. This commit adds support for
the parsing of the v1 and v2 PROXY protocol header but does not yet
parse the TLV data in the v2 header.

Check if a PROXY protocol header was sent by the client at the
beginning of the connection. If so, parse it, store the extracted
informations and replace the proxy address with the real client
address. Checking for the header is done only if the packet has an
invalid length so that there's minimal overhead for the common case.

Also add a small document file showing how to configure HAProxy to
test the feature and a unit test for the parsing.

Change-Id: I9766c68feaba71279f26c3310eee4b5ca215e6ac
Signed-off-by: Ralf Lici <ralf@mandelbit.com>