dco-linux: pass remoteaddr only for UDP peers
TCP connected sockets do not need any remote addr
because the destination is fixed.
For this reason we can avoid sending the remote addr
along the peer-new dco call.

This change is important because the new 'ovpn' kernel
module is stricter when it comes to accepting netlink
messages and will reject calls with TCP sockets if
a remote address is specified.

Signed-off-by: Antonio Quartulli <antonio@mandelbit.com>