redirect-gateway: redirect the traffic through the tun only if redirect-gateway is of the same type of the tun's addresses.
Fixing ifconfig push-reply bug: if remote is switched and the new tun has a different address family, the previous ifconfig
options remains to be assigned to the new tun.

Adding a check in do_init_route_ipv6_list() to add the default routes towards the tun only if the tun has IPv6 addresses.

Change-Id: Ib3458a9ed2eb38e00184c4a92659b83b97fe476c