sitnl: open rtnl socket once and keep it open
Currently, the sitnl interface opens and closes the rtnl socket for
every message. To reduce per-message overhead and prepare for the
upcoming self-destruction logic for DCO interfaces, open the socket once
on the first message and keep it open until program termination.

ovpn tun interfaces will rely on this socket as a lifeline for the
userspace program, so it must remain open for the duration of the
program's lifetime.

Change-Id: I9c6742dc5cd3759f576a642b8c91038b0c510372
Signed-off-by: Ralf Lici <ralf@mandelbit.com>