dco_linux: create ovpn interfaces as self-destroying
When creating an interface, pass a flag to ovpn to indicate that the
interface should be automatically deleted when openvpn terminates,
regardless of the termination condition.

With this flag set, the kernel module monitors the associated netlink
socket and tears down the interface once the socket is closed. This
ensures that even if userspace crashes, the interface is still removed
when the process exits and the socket is released.

Change-Id: I1b4cfa12c0298624e20b60275cf6bc8cc0703e80
Signed-off-by: Ralf Lici <ralf@mandelbit.com>