dco_linux: pass PID to ovpn when creating a self-destroying interface
To signal that the interface should be deleted when the userspace
program terminates (regardless of the termination condition) pass the
program's PID to ovpn.

The kernel module will monitor the associated rtnl socket, and once it's
closed, automatically tear down the interface. Even if userspace crashes
the kernel will still close the socket when the process is terminated.

Change-Id: I1b4cfa12c0298624e20b60275cf6bc8cc0703e80
Signed-off-by: Ralf Lici <ralf@mandelbit.com>