Properly log peer-id in P2P mode
As of now OpenVPN prints o->peer_id in the deferred
options log message, however in P2P mode that field
is currently always 0 as it gets populated only when
a peer-id is pushed by a server (which doesn't happen in P2P).

Fix this by populate o->peer_id before loggin it.

Change-Id: I80aed8a91259374d4b2b2155a8f47e286f2def75
Signed-off-by: Gianmarco De Gregori <gianmarco@mandelbit.com>